Check token

+++ Gateway/blueprints/ABService.py
@@ -9,6 +9,13 @@ URL = "http://127.0.0.1:8090/api/bookcollection/"
 
 @abService.route('/books')
 def get_all_books():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     param = {}
     genre = request.args.get("genre")
     if genre is not None:
@@ -28,6 +35,13 @@ def get_all_books():
 
 @abService.route('/books/<ISBN>')
 def get_book_by_isbn(ISBN):
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     verbose = request.args.get("verbose")
     r = requests.get(parse.urljoin(URL, "books/" + ISBN), params={'verbose': verbose})
     return r.content,r.status_code
@@ -35,6 +49,13 @@ def get_book_by_isbn(ISBN):
 
 @abService.route('/books/<ISBN>', methods=['DELETE'])
 def delete_book_by_isbn(ISBN):
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     r = requests.delete(parse.urljoin(URL, "books/" + ISBN))
     return r.content,r.status_code
 
@@ -42,6 +63,13 @@ def delete_book_by_isbn(ISBN):
 # TODO: json schema validator
 @abService.route('/books', methods=['POST'])
 def add_books():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     json = request.json
     r = requests.post(parse.urljoin(URL, "books"), json=json)
     return r.content,r.status_code
@@ -50,6 +78,13 @@ def add_books():
 # TODO: redirect to error when name or match missing
 @abService.route('/authors')
 def get_authors():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     param = {}
     name = request.args.get("name")
     param.update({"name": name})
@@ -62,6 +97,13 @@ def get_authors():
 # TODO: json schema validator for authors
 @abService.route('/authors', methods=['POST'])
 def add_author():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     json = request.json
     r = requests.post(parse.urljoin(URL, "authors"), json=json)
     return r.content,r.status_code
@@ -70,36 +112,78 @@ def add_author():
 # TODO: hateoas in main app for authors
 @abService.route('/authors/<ID>')
 def get_author_by_id(ID):
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     r = requests.get(parse.urljoin(URL, "authors/" + ID))
     return r.content,r.status_code
 
 
 @abService.route('/authors/<ID>', methods=['DELETE'])
 def delete_author_by_id(ID):
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     r = requests.delete(parse.urljoin(URL, "authors/" + ID))
     return r.content,r.status_code
 
 
 @abService.route('/', methods=['OPTIONS'])
 def get_all_methods():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     r = requests.options(URL)
     return r.content,r.status_code
 
 
 @abService.route('/books', methods=['OPTIONS'])
 def get_methods_books():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     r = requests.options(parse.urljoin(URL, "books"))
     return r.content,r.status_code
 
 
 @abService.route('/authors', methods=['OPTIONS'])
 def get_methods_authors():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     r = requests.options(parse.urljoin(URL, "authors"))
     return r.content,r.status_code
 
 
 @abService.route('/books/<ISBN>/authors', methods=['POST'])
 def add_authors_to_books(ISBN):
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     json = request.json
     r = requests.post(parse.urljoin(URL, "books/" + ISBN + "/authors"), json=json)
     return r.content,r.status_code
@@ -107,6 +191,13 @@ def add_authors_to_books(ISBN):
 
 @abService.route('/books/stockChange', methods=['POST'])
 def stock_endpoint():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     json = request.json
     r = requests.post(parse.urljoin(URL, "books/stockChange"), json=json)
     return r.content,r.status_code
