+++ Gateway/blueprints/cartService.py
@@ -13,6 +13,13 @@ URL = "http://127.0.0.1:8093/"
 
 @cart.route('/cart')
 def get_cart_by_client_id():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     client_id = request.args['clientid']
     r = requests.get(parse.urljoin(URL, "cart?clientid=" + client_id))
     return r.content,r.status_code
@@ -20,6 +27,13 @@ def get_cart_by_client_id():
 
 @cart.route('/cart', methods=['POST'])
 def add_item_in_cart():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     client_id = request.args['clientid']
     json = request.json
     r = requests.post(parse.urljoin(URL, "cart?clientid=" + client_id), json=json)
@@ -28,6 +42,13 @@ def add_item_in_cart():
 
 @cart.route('/cart', methods=['DELETE'])
 def delete_all_cart():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     client_id = request.args['clientid']
     r = requests.delete(parse.urljoin(URL, "cart?clientid=" + client_id))
     return r.content,r.status_code
@@ -35,6 +56,13 @@ def delete_all_cart():
 
 @cart.route('/addCart', methods=['PUT'])
 def add_quantity():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     client_id = request.args['clientid']
     isbn = request.args['ISBN']
     r = requests.put(parse.urljoin(URL, "addCart?clientid=" + client_id + "&ISBN=" + isbn), json=None)
@@ -43,6 +71,13 @@ def add_quantity():
 
 @cart.route('/downCart', methods=['PUT'])
 def down_quantity():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     client_id = request.args['clientid']
     isbn = request.args['ISBN']
     r = requests.put(parse.urljoin(URL, "downCart?clientid=" + client_id + "&ISBN=" + isbn), json=None)
@@ -51,6 +86,13 @@ def down_quantity():
 
 @cart.route('/removeItem', methods=['DELETE'])
 def remove_item():
+
+    head = request.headers
+    r = requests.get(parse.urljoin("http://127.0.0.1:5000/", "check-token"), headers=head)
+
+    if r.status_code != 200:
+        return r.content, r.status_code
+
     client_id = request.args['clientid']
     isbn = request.args['ISBN']
     r = requests.delete(parse.urljoin(URL, "removeItem?clientid=" + client_id + "&ISBN=" + isbn), json=None)
